update vendor module

+++ internal/domain/repositories/vendor_repository.go
@@ -8,7 +8,7 @@ import (
 )
 
 type VendorRepository interface {
-	GetAllVendors(ctx context.Context, vendorName string, city string, limit int, offset int, sortBy string, sortOrder string) ([]*entities.Vendor, int64, error)
+	GetAllVendors(ctx context.Context, vendorName string, city string, limit int, offset int, sortBy string, sortOrder string) ([]*entities.Vendor, int, error)
 	GetVendorByID(ctx context.Context, vendorID uuid.UUID) (*entities.Vendor, error)
 	GetVendorsResourcesByVendorID(ctx context.Context, vendorID uuid.UUID) ([]*entities.Resource, error)
 	// GetVendorReviews(ctx context.Context, vendorID uuid.UUID) ([]*entities.Review, error)
